+++ src/user/user.service.ts
@@ -7,6 +7,7 @@ import { DbUser } from "./user.entity";
 import { UpdateUserDto } from "./dtos/update-user.dto";
 import { UpdateUserProfileInterestsDto } from "./dtos/update-user-interests.dto";
 import { UpdateUserEmailPreferencesDto } from "./dtos/update-user-email-prefs.dto";
+import { UserOnboardingDto } from "../auth/dtos/user-onboarding.dto";
 
 @Injectable()
 export class UserService {
@@ -64,7 +65,7 @@ export class UserService {
 
   async checkAddUser (user: User): Promise<DbUser> {
     const {
-      user_metadata: { user_name, email },
+      user_metadata: { user_name, email, name },
       identities,
     } = user;
     const github = identities!.filter(identity => identity.provider === "github")[0];
@@ -76,7 +77,8 @@ export class UserService {
       // create new user
       const newUser = this.userRepository.create({
         id,
-        is_open_sauced_member: false,
+        name: name as string,
+        is_open_sauced_member: true,
         login: user_name as string,
         email: email as string,
         created_at: (new Date),
@@ -103,6 +105,8 @@ export class UserService {
         location: user.location ?? "",
         display_local_time: !!user.display_local_time,
         timezone: user.timezone,
+        github_sponsors_url: user.github_sponsors_url ?? "",
+        linkedin_url: user.linkedin_url ?? "",
       });
 
       return this.findOneById(id);
@@ -111,11 +115,16 @@ export class UserService {
     }
   }
 
-  async updateOnboarding (id: number) {
+  async updateOnboarding (id: number, user: UserOnboardingDto) {
     try {
       await this.findOneById(id);
 
-      await this.userRepository.update(id, { is_onboarded: true, is_waitlisted: false });
+      await this.userRepository.update(id, {
+        is_onboarded: true,
+        is_waitlisted: false,
+        timezone: user.timezone,
+        interests: user.interests.join(","),
+      });
     } catch (e) {
       throw new NotFoundException("Unable to update user onboarding status");
     }